Move implementation of StyleToken to separate files (#481)
